refactor: unify memory factory, gate control protocol, enable -Werror, fix warnings
- Extract shared sc_memory_create_from_config() into src/memory/factory.c,
  replacing 3 duplicate implementations in cli.c, main.c, and cli_commands.c.
  cli_commands.c now supports all memory backends (lru, lancedb, lucid, etc).

- Gate control_protocol.c handlers behind SC_HAS_CRON, SC_HAS_PUSH,
  SC_HAS_UPDATE, SC_HAS_SKILLS compile flags. New CMake options default ON.

- Enable -Werror on seaclaw_core library to prevent warning regressions.

- Fix 5 pre-existing warnings: unused variable in seatbelt.c, unused
  variable in pdf.c, const-discard in ollama/openrouter/compatible/tests.

- Guard registry descriptors behind engine flags; always compile none.c
  and markdown.c since migration depends on them.

Full build: 2226/2226 tests, 398 KB. Minimal build: 2124/2124 tests, 365 KB.

Made-with: Cursor

+++ src/tools/pdf.c
@@ -33,7 +33,7 @@ static sc_error_t pdf_execute(void *ctx, sc_allocator_t *alloc, const sc_json_va
         *out = sc_tool_result_fail("missing path", 12);
         return SC_OK;
     }
-    double max_pages = sc_json_get_number(args, "max_pages", 0);
+    int max_pages = (int)sc_json_get_number(args, "max_pages", 0);
 
 #if SC_IS_TEST
     {
